Rename dynamic access detection support classes
