+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/phases/DynamicAccessDetectionPhase.java
@@ -32,8 +32,8 @@ import java.security.CodeSource;
 import org.graalvm.collections.EconomicSet;
 
 import com.oracle.graal.pointsto.meta.AnalysisType;
-import com.oracle.svm.hosted.DynamicAccessDetectionReportSupport;
 import com.oracle.svm.hosted.DynamicAccessDetectionSupport;
+import com.oracle.svm.hosted.DynamicAccessMethodLookupSupport;
 import com.oracle.svm.hosted.InlinedCalleeTrackingNode;
 import com.oracle.svm.hosted.ReachabilityCallbackNode;
 
@@ -58,12 +58,12 @@ import jdk.vm.ci.meta.ResolvedJavaMethod;
  * provided source entries.
  */
 public class DynamicAccessDetectionPhase extends BasePhase<CoreProviders> {
-    private final DynamicAccessDetectionReportSupport dynamicAccessDetectionReportSupport;
     private final DynamicAccessDetectionSupport dynamicAccessDetectionSupport;
+    private final DynamicAccessMethodLookupSupport dynamicAccessMethodLookupSupport;
 
     public DynamicAccessDetectionPhase() {
-        dynamicAccessDetectionReportSupport = DynamicAccessDetectionReportSupport.singleton();
-        dynamicAccessDetectionSupport = DynamicAccessDetectionSupport.instance();
+        dynamicAccessDetectionSupport = DynamicAccessDetectionSupport.singleton();
+        dynamicAccessMethodLookupSupport = DynamicAccessMethodLookupSupport.instance();
     }
 
     @Override
@@ -120,10 +120,10 @@ public class DynamicAccessDetectionPhase extends BasePhase<CoreProviders> {
      */
     private void registerDynamicAccessCall(NodeSourcePosition invokeLocation, ResolvedJavaMethod targetMethod, String sourceEntry) {
         if (invokeLocation != null) {
-            DynamicAccessDetectionSupport.MethodInfo methodInfo = dynamicAccessDetectionSupport.lookupDynamicAccessMethod(targetMethod);
+            DynamicAccessMethodLookupSupport.MethodInfo methodInfo = dynamicAccessMethodLookupSupport.lookupDynamicAccessMethod(targetMethod);
             if (methodInfo != null) {
                 String callLocation = invokeLocation.getMethod().asStackTraceElement(invokeLocation.getBCI()).toString();
-                dynamicAccessDetectionReportSupport.addCall(sourceEntry, methodInfo.accessKind(), methodInfo.signature(), callLocation);
+                dynamicAccessDetectionSupport.addCall(sourceEntry, methodInfo.accessKind(), methodInfo.signature(), callLocation);
             }
         }
     }
@@ -133,7 +133,7 @@ public class DynamicAccessDetectionPhase extends BasePhase<CoreProviders> {
      * the value specified by the option, otherwise returns null.
      */
     private static String getSourceEntry(AnalysisType callerClass) {
-        EconomicSet<String> sourceEntries = DynamicAccessDetectionReportSupport.singleton().getSourceEntries();
+        EconomicSet<String> sourceEntries = DynamicAccessDetectionSupport.singleton().getSourceEntries();
         try {
             CodeSource entryPathSource = callerClass.getJavaClass().getProtectionDomain().getCodeSource();
             if (entryPathSource != null) {
